+++ database/migrations/2023_04_18_150103_create_cart_items_table.php
@@ -10,9 +10,9 @@ class CreateCartItemsTable extends Migration
     {
         Schema::create('cart_items', function (Blueprint $table) {
             $table->id();
-            $table->unsignedInteger('cart_id')->comment("Id корзини");
-            $table->unsignedInteger('product_id')->comment("Id товару");
-            $table->unsignedInteger('quantity')->comment("Кількість");
+            $table->unsignedBigInteger('cart_id')->comment("Id корзини");
+            $table->unsignedBigInteger('product_id')->comment("Id товару");
+            $table->unsignedBigInteger('quantity')->comment("Кількість");
             $table->decimal('price')->comment("Вартість");
             $table->decimal('pricediscount')->nullable()->comment("Знижка");
             $table->boolean('active')->unsigned()->default(1);
