+++ database/migrations/2023_04_18_150103_create_wish_items_table.php
@@ -10,8 +10,8 @@ class CreateWishItemsTable extends Migration
     {
         Schema::create('wish_items', function (Blueprint $table) {
             $table->id();
-            $table->unsignedInteger('user_id')->comment("Id користувача");
-            $table->unsignedInteger('product_id')->comment("Id товару");
+            $table->unsignedBigInteger('user_id')->comment("Id користувача");
+            $table->unsignedBigInteger('product_id')->comment("Id товару");
             $table->decimal('price')->comment("Вартість");
             $table->boolean('active')->unsigned()->default(1);
             $table->softDeletes();
