+++ database/migrations/2023_04_19_143010_create_orders_table.php
@@ -10,20 +10,20 @@ class CreateOrdersTable extends Migration
     {
         Schema::create('orders', function (Blueprint $table) {
             $table->id();
-            $table->unsignedInteger('user_id')->comment("Id користувача");
+            $table->unsignedBigInteger('user_id')->comment("Id користувача");
             $table->decimal('sum_order',12,2)->comment("сума замовлення");
-            $table->unsignedInteger('card')->nullable()->comment("карта"); // ??
-            $table->unsignedInteger('cart_id')->comment("Id корзини");
-            $table->unsignedInteger('kind_payment_id')->comment("Id виду оплати");
+            $table->unsignedBigInteger('card')->nullable()->comment("карта"); // ??
+            $table->unsignedBigInteger('cart_id')->comment("Id корзини");
+            $table->unsignedBigInteger('kind_payment_id')->comment("Id виду оплати");
             $table->text('comment')->nullable()->comment("коментар замовлення");
-            $table->unsignedInteger('delivery_id')->comment("Id виду доставки");
-            $table->unsignedInteger('region_id')->comment("Id області");
-            $table->unsignedInteger('city_id')->comment("Id міста");
+            $table->unsignedBigInteger('delivery_id')->comment("Id виду доставки");
+            $table->unsignedBigInteger('region_id')->comment("Id області");
+            $table->unsignedBigInteger('city_id')->comment("Id міста");
             $table->string('address', 255)->comment("Адреса");
-            $table->unsignedInteger('newpost_id')->nullable()->comment("Id НП");
+            $table->unsignedBigInteger('newpost_id')->nullable()->comment("Id НП");
             $table->string('promocode', 255)->nullable()->comment("Промокод");
             $table->decimal('pricedelivery', 12, 2)->nullable()->comment("Вартсіть доставки");
-            $table->unsignedInteger('status_order_id')->comment("Статус замовлення");
+            $table->unsignedBigInteger('status_order_id')->comment("Статус замовлення");
             $table->boolean('active')->default(true);
             $table->softDeletes();
             $table->timestamps();
