+++ database/migrations/2024_04_05_172833_create_forum_sub_categories_table.php
@@ -16,7 +16,7 @@ return new class extends Migration
         Schema::create('forum_sub_categories', function (Blueprint $table) {
             $table->id();
             $table->string('name')->comment('Назва підкатегорії форума');
-            $table->unsignedInteger('forum_category_id')->comment('Id категорії форума');
+            $table->unsignedBigInteger('forum_category_id')->comment('Id категорії форума');
             $table->softDeletes();
             $table->timestamps();
